refactor: reorganize RideResponse and RideStopRequest classes into appropriate packages
